- 修复 ISelect Sum/First 子查询时，若子查询实体类与主查询一样时，导致的 bug；

+++ FreeSql.Tests/FreeSql.Tests.Provider.MySqlConnector/MySqlConnector/Curd/MySqlSelectTest.cs
@@ -884,12 +884,12 @@ namespace FreeSql.Tests.MySqlConnector
             var subquery = select.ToSql(a => new
             {
                 all = a,
-                count = select.Sum(b => b.Id)
+                count = (long)select.Sum(b => b.Id)
             });
             var subqueryList = select.ToList(a => new
             {
                 all = a,
-                count = select.Sum(b => b.Id)
+                count = (long)select.Sum(b => b.Id)
             });
         }
         [Fact]
