+++ FreeSql.Tests/FreeSql.Tests/MySql/Curd/MySqlSelectTest.cs
@@ -874,12 +874,12 @@ namespace FreeSql.Tests.MySql
             var subquery = select.ToSql(a => new
             {
                 all = a,
-                count = select.Sum(b => b.Id)
+                count = (long)select.Sum(b => b.Id)
             });
             var subqueryList = select.ToList(a => new
             {
                 all = a,
-                count = select.Sum(b => b.Id)
+                count = (long)select.Sum(b => b.Id)
             });
         }
         [Fact]
